added yagmail and utility functions to send email and generate passwords
